# contract/openapi.yaml
...
openapi: 3.1.0
info:
  title: Certa API
  version: 0.1.0
  description: Flight operations & compliance logbook API. All quantities are SI; all instants are UTC ISO-8601. Org-scoped routes live under /api/v1/orgs/{orgId}.
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: better-auth.session_token
    bearer:
      type: http
      scheme: bearer
...
paths:
...
  /api/v1/orgs:
    post:
      summary: Create an additional organization owned by the caller
      tags:
        - organizations
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 200
                timezone:
                  type: string
                  minLength: 1
                  maxLength: 64
                defaultJurisdiction:
                  default: US-FAA-Part107
                  type: string
                  pattern: ^[A-Z]{2}-[A-Za-z0-9]+(-[A-Za-z0-9]+)*$
                units:
                  type: object
                  properties:
                    length:
                      type: string
                      enum:
                        - ft
                        - m
                    speed:
                      type: string
                      enum:
                        - mph
                        - kph
                        - kt
                        - mps
                    mass:
                      type: string
                      enum:
                        - lb
                        - kg
                        - g
                    temperature:
                      type: string
                      enum:
                        - F
                        - C
                  required:
                    - length
                    - speed
                    - mass
                    - temperature
              required:
                - name
                - timezone
      responses:
        "201":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  organization:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                        pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                      name:
                        type: string
                        minLength: 1
                        maxLength: 200
                      slug:
                        type: string
                        pattern: ^[a-z0-9-]{2,64}$
                      planTier:
                        type: string
                        enum:
                          - solo
                          - team
                          - enterprise
                          - self_hosted
                      defaultJurisdiction:
                        type: string
                        pattern: ^[A-Z]{2}-[A-Za-z0-9]+(-[A-Za-z0-9]+)*$
                      timezone:
                        type: string
                        minLength: 1
                        maxLength: 64
                      units:
                        type: object
                        properties:
                          length:
                            type: string
                            enum:
                              - ft
                              - m
                          speed:
                            type: string
                            enum:
                              - mph
                              - kph
                              - kt
                              - mps
                          mass:
                            type: string
                            enum:
                              - lb
                              - kg
                              - g
                          temperature:
                            type: string
                            enum:
                              - F
                              - C
                        required:
                          - length
                          - speed
                          - mass
                          - temperature
                        additionalProperties: false
                      createdAt:
                        type: string
                        format: date-time
                        pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d:[0-5]\d(?:\.\d+)?(?:Z|([+-](?:[01]\d|2[0-3]):[0-5]\d)))$
                      updatedAt:
                        type: string
                        format: date-time
                        pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d:[0-5]\d(?:\.\d+)?(?:Z|([+-](?:[01]\d|2[0-3]):[0-5]\d)))$
                    required:
                      - id
                      - name
                      - slug
                      - planTier
                      - defaultJurisdiction
                      - timezone
                      - units
                      - createdAt
                      - updatedAt
                    additionalProperties: false
                  membership:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                        pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                      orgId:
                        type: string
                        format: uuid
                        pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                      createdAt:
                        type: string
                        format: date-time
                        pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d:[0-5]\d(?:\.\d+)?(?:Z|([+-](?:[01]\d|2[0-3]):[0-5]\d)))$
                      updatedAt:
                        type: string
                        format: date-time
                        pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d:[0-5]\d(?:\.\d+)?(?:Z|([+-](?:[01]\d|2[0-3]):[0-5]\d)))$
                      createdBy:
                        anyOf:
                          - type: string
                            format: uuid
                            pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                          - type: "null"
                      deletedAt:
                        anyOf:
                          - type: string
                            format: date-time
                            pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d:[0-5]\d(?:\.\d+)?(?:Z|([+-](?:[01]\d|2[0-3]):[0-5]\d)))$
                          - type: "null"
                      userId:
                        type: string
                        format: uuid
                        pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                      role:
                        type: string
                        enum:
                          - owner
                          - admin
                          - chief_pilot
                          - pilot
                          - maintenance_tech
                          - viewer
                          - auditor
                      expiresAt:
                        anyOf:
                          - type: string
                            format: date-time
                            pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d:[0-5]\d(?:\.\d+)?(?:Z|([+-](?:[01]\d|2[0-3]):[0-5]\d)))$
                          - type: "null"
                      auditorScope:
                        anyOf:
                          - type: object
                            properties:
                              from:
                                anyOf:
                                  - type: string
                                    format: date
                                    pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))$
                                  - type: "null"
                              to:
                                anyOf:
                                  - type: string
                                    format: date
                                    pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))$
                                  - type: "null"
                              pilotIds:
                                anyOf:
                                  - type: array
                                    items:
                                      type: string
                                      format: uuid
                                      pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                                  - type: "null"
                              aircraftIds:
                                anyOf:
                                  - type: array
                                    items:
                                      type: string
                                      format: uuid
                                      pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                                  - type: "null"
                            required:
                              - from
                              - to
                              - pilotIds
                              - aircraftIds
                            additionalProperties: false
                          - type: "null"
                    required:
                      - id
                      - orgId
                      - createdAt
                      - updatedAt
                      - createdBy
                      - deletedAt
                      - userId
                      - role
                      - expiresAt
                      - auditorScope
                    additionalProperties: false
                  permissions:
                    type: array
                    items:
                      type: string
                      enum:
                        - org.read
                        - org.update
                        - org.billing
                        - members.read
                        - members.manage
                        - rules.read
                        - rules.override
                        - audit.read
                        - pilots.read
                        - pilots.manage
                        - credentials.read
                        - credentials.manage
                        - aircraft.read
                        - aircraft.manage
                        - aircraft.ground
                        - aircraft.override_grounding
                        - batteries.read
                        - batteries.manage
                        - components.manage
                        - maintenance.read
                        - maintenance.manage
                        - flights.read
                        - flights.log
                        - flights.edit
                        - missions.read
                        - missions.manage
                        - checklists.read
                        - checklists.manage_templates
                        - checklists.run
                        - incidents.read
                        - incidents.report
                        - incidents.manage
                        - exports.create
                        - integrity.verify
                        - integrations.manage
                required:
                  - organization
                  - membership
                  - permissions
                additionalProperties: false
...
security:
  - session: []
  - bearer: []